限流逻辑完善，增加密码抽奖

+++ server/src/main/java/com/lotterysystem/server/controller/LotteryController.java
@@ -1,11 +1,14 @@
 package com.lotterysystem.server.controller;
 
 import cn.hutool.json.JSONObject;
+import com.lotterysystem.gateway.SentinelConfig.LotteryActionLimiter;
 import com.lotterysystem.gateway.util.UserContext;
+import com.lotterysystem.server.constant.AuthStatue;
 import com.lotterysystem.server.constant.ResultStatue;
 import com.lotterysystem.server.pojo.dto.LotteryDTO;
 import com.lotterysystem.server.pojo.dto.Result;
 import com.lotterysystem.server.service.LotteryService;
+import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
@@ -24,17 +27,24 @@ public class LotteryController {
     final LotteryService lotteryService;
 
     @PostMapping
+    @Schema(description = "新增抽奖活动")
     public Result addLottery(@RequestBody LotteryDTO lotteryDTO) throws Exception {
+        if(!chooseLimiter())
+            return defaultFallback();
         return lotteryService.addLottery(lotteryDTO);
     }
-    //获取所有的抽奖信息
+
+    //获取所有的抽奖信息（只有抽奖）
     @GetMapping("/your")
     public JSONObject getAllLottery(@RequestParam int page) {
         return lotteryService.getAllLottery(page);
     }
 
     @PutMapping
+    @Schema(description = "更新抽奖，比较吃性能")
     public Result updateLottery(@RequestBody LotteryDTO lotteryDTO) throws Exception {
+        if(!chooseLimiter())
+            return defaultFallback();
         return lotteryService.updateLottery(lotteryDTO);
     }
 
@@ -44,12 +54,18 @@ public class LotteryController {
     }
 
     @DeleteMapping
+    @Schema(description = "会删除所有数据，包括任务与结果")
     public Result deleteLottery(@RequestParam("id") Long id) throws Exception {
+        if(!chooseLimiter())
+            return defaultFallback();
         return lotteryService.deleteLottery(id);
     }
 
     @PostMapping("/stop")
+    @Schema(description = "提前停止抽奖活动,比较吃性能")
     public Result stopLottery(@RequestParam("id") Long id) {
+        if(!chooseLimiter())
+            return defaultFallback();
         return lotteryService.stopLottery(id);
     }
 
@@ -57,9 +73,23 @@ public class LotteryController {
     public Result getOverview() {
         return new Result(ResultStatue.SUCCESS,"查询成功!",lotteryService.getOverView(UserContext.getId())) ;
     }
+
     @GetMapping("/myall")
     public Result getAllInfo(){
         return new Result(ResultStatue.SUCCESS,"查询成功!",lotteryService.getAllInfo(UserContext.getId())) ;
+    }
+
+    public boolean chooseLimiter(){
+        Long userId = UserContext.getId();
+        Integer auth = UserContext.getAuth();
+        if(auth == AuthStatue.USER.getCode())
+            return LotteryActionLimiter.tryUserAccess(userId);
+        else
+            return LotteryActionLimiter.tryAdminAccess(userId);
+    }
+
+    public Result defaultFallback() {
+        return new Result(ResultStatue.SC_SERVICE_UNAVAILABLE,"您操作得太快，请待会再重试！",null);
     }
 
 }
